git-svn-id: svn://123.57.173.36/meijia-parent@679 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-dao/src/main/java/com/simi/po/dao/partners/PartnerServicePricesMapper.java
@@ -8,9 +8,9 @@ import com.simi.po.model.partners.PartnerServiceType;
 public interface PartnerServicePricesMapper {
     int deleteByPrimaryKey(Long servicePriceId);
 
-    int insert(PartnerServicePrices record);
+    Long insert(PartnerServicePrices record);
 
-    int insertSelective(PartnerServicePrices record);
+    Long insertSelective(PartnerServicePrices record);
 
     PartnerServicePrices selectByPrimaryKey(Long servicePriceId);
